Move sanitize code to version
