+++ mepo.d/command/status/status.py
@@ -5,7 +5,7 @@ import atexit
 
 from state.state import MepoState
 from repository.git import GitRepository
-from utilities.version import version_to_string
+from utilities.version import version_to_string, sanitize_version_string
 from utilities import colors
 
 def run(args):
@@ -55,52 +55,3 @@ def print_status(allcomps, result):
         if (output):
             for line in output.split('\n'):
                 print('   |', line.rstrip())
-
-def sanitize_version_string(orig,curr,git):
-    '''
-    This routine tries to figure out if two tags are the same.
-
-    The issue is that git sometimes returns the "wrong" tag in
-    the mepo sense. mepo might have checked out tag v1.0.0 but
-    if that commit is also tagged with foo, then sometimes mepo
-    will say that things have changed because it thinks it's on
-    foo.
-    '''
-
-    # The trick below only works on tags (I think), so
-    # if not a tag, just do nothing for now
-    is_tag = '(t)'
-
-    # We pass in space-delimited strings that are:
-    #  'type version dh'
-    # So let's split into lists...
-    orig_list = orig.split()
-    curr_list = curr.split()
-
-    # Then pull the types...
-    orig_type = orig_list[0]
-    curr_type = curr_list[0]
-
-    # Now if a type...
-    if orig_type == is_tag and curr_type == is_tag:
-
-        # Pull out the version string...
-        orig_ver = orig_list[1]
-        curr_ver = curr_list[1]
-
-        # Use rev-list to get the hash of the tag...
-        orig_rev = git.rev_list(orig_ver)
-        curr_rev = git.rev_list(curr_ver)
-
-        # If they are identical...
-        if orig_rev == curr_rev:
-
-            # Replace the curr version with the original to make
-            # mepo happy...
-            curr_list[curr_list.index(curr_ver)] = orig_ver
-
-            # And then remake the curr string
-            curr = ' '.join(curr_list)
-
-    # And return curr
-    return curr
